+++ src/Musicus.Abstractions/Services/IMusicService.cs
@@ -14,7 +14,7 @@ namespace Musicus.Abstractions.Services
 		Task<bool> NextAsync(string url);
 		Task<IList<ISearchResult>> SearchAsync(string keyword);
 		Task<IMusicServiceStatus> GetStatusAsync();
-		Task SetVolumeAsync(int volume);
+		Task SetVolumeAsync(float volume);
 		Task<float> GetVolumeAsync();
 	}
 }
