+++ src/Musicus/Models/Track.cs
@@ -5,7 +5,7 @@ namespace Musicus.Models
 	public class Track
 	{
 		public string TrackId { get; set; }
-		public int TrackLength { get; set; }
+		public double TrackLength { get; set; }
 		public string Artist { get; set; }
 		public string Description { get; set; }
 		public string Url { get; set; }
